VolonterskiSati v1.0
Account list func., Archiving members func, optimized views

+++ members.py
@@ -4,9 +4,10 @@ from flask import (
 from werkzeug.security import generate_password_hash
 from werkzeug.exceptions import HTTPException
 
-from auth import login_required
+from auth import login_required, savjetnik_required, admin_required
 
 import access_levels
+import Utilities
 
 from DatabaseController import DatabaseController, get_date_object
 import DatabaseTables
@@ -16,6 +17,8 @@ members_bp = Blueprint('members', __name__, url_prefix='/members')
 
 
 @members_bp.route('/add', methods=('GET', 'POST'))
+@login_required
+@savjetnik_required
 def add_member():
     if request.method == 'POST':
         name = request.form['name']
@@ -27,33 +30,38 @@ def add_member():
         membership_start = request.form['membership']
         card_id = request.form['idcard']
         email = request.form['email']
-        section = request.form['section']
+        try:
+            section = request.form['section']
+        except HTTPException:
+            section = session['section']
 
         db = DatabaseController()
         error = None
 
         if not name:
             error = 'Username is required.'
-        elif not last_name:
+        if not last_name:
             error = 'Password is required.'
-        elif not nickname:
+        if not nickname:
             nickname = '-'
-        elif not oib:
+        if not oib:
             error = 'Oib value is required.'
-        elif not phone_number:
+        if not phone_number:
             error = 'Phone number is required.'
-        elif not date_of_birth:
+        if not date_of_birth:
             error = 'Date of birth is required.'
-        elif not membership_start:
+        if not membership_start:
             error = 'Membership start date is required.'
-        elif not card_id:
+        if not card_id:
             error = 'Card id is required.'
-        elif not email:
+        if not email:
             error = 'Email is required.'
-        elif not section or section == 'Izaberi sekciju':
+        if not section or section == 'Izaberi sekciju':
             error = 'Sekcija value is required.'
-        elif db.account_exists(card_id):
-            error = 'User card id {} is already registered.'.format("%s %s" % (name, last_name))
+        if db.member_exists(card_id.strip()):
+            registered_member = db.get_row(DatabaseTables.CLAN, 'broj_iskaznice', card_id.strip())
+            name, last_name = registered_member[1], registered_member[2]
+            error = 'Broj iskaznice već postoji i glasi na ime {0} {1}'.format(name, last_name)
 
         if error is None:
             date_of_birth = get_date_object(date_of_birth)
@@ -81,7 +89,7 @@ def list_members():
     members_list = {}
     for member in members:
         if db.is_member_active(member[0]):
-            members_list[member[0]] = member[1:-1]
+            members_list[member[0]] = member[1:-2]
 
     if session["access_level"] >= access_levels.SAVJETNIK:
         sorted_list = sorted(members_list.items(), key=lambda x: x[1][1])  # Sort po prezimenu ako je unutar sekcije
@@ -95,6 +103,8 @@ def list_members():
 
 
 @members_bp.route('/edit/<member_id>', methods=['GET', 'POST'])
+@login_required
+@savjetnik_required
 def edit_member(member_id):
     db = DatabaseController()
     member = db.get_row(DatabaseTables.CLAN, 'id', member_id)
@@ -111,33 +121,35 @@ def edit_member(member_id):
         try:
             section = request.form['section']
         except HTTPException:
-            section = member[9]
+            section = member[-3]
 
         db = DatabaseController()
         error = None
 
         if not name:
             error = 'Username is required.'
-        elif not last_name:
+        if not last_name:
             error = 'Password is required.'
-        elif not nickname:
+        if not nickname:
             nickname = '-'
-        elif not oib:
+        if not oib:
             error = 'Oib value is required.'
-        elif not phone_number:
+        if not phone_number:
             error = 'Phone number is required.'
-        elif not date_of_birth:
+        if not date_of_birth:
             error = 'Date of birth is required.'
-        elif not membership_start:
+        if not membership_start:
             error = 'Membership start date is required.'
-        elif not card_id:
+        if not card_id:
             error = 'Card id is required.'
-        elif not email:
+        if not email:
             error = 'Email is required.'
-        elif not section or section == 'Izaberi sekciju':
+        if not section or section == 'Izaberi sekciju':
             error = 'Sekcija value is required.'
-        elif card_id != member[8] and db.account_exists(card_id):
-            error = 'User card id {} is already registered.'.format("%s %s" % (name, last_name))
+        if card_id != member[8] and db.member_exists(card_id.strip()):
+            registered_member = db.get_row(DatabaseTables.CLAN, 'broj_iskaznice', card_id.strip())
+            name, last_name = registered_member[1], registered_member[2]
+            error = 'Broj iskaznice već postoji i glasi na ime {0} {1}'.format(name, last_name)
 
         if error is None:
             date_of_birth = get_date_object(date_of_birth)
@@ -150,10 +162,12 @@ def edit_member(member_id):
 
         flash(error, 'danger')
 
-    return render_template('/members/edit.html', member=member[1:], member_id=member[0])
+    return render_template('/members/edit.html', member=member[1:-1], member_id=member[0], sections=Utilities.sections)
 
 
 @members_bp.route('/remove/<member_id>', methods=['POST'])
+@savjetnik_required
+@login_required
 def remove_member(member_id):
 
     if request.method == 'POST':
@@ -163,8 +177,57 @@ def remove_member(member_id):
             error = 'Neuspješno brisanje. Zapis ne postoji u bazi.'
             flash(error, 'danger')
         else:
+            db.deactivate_member(member_id)
+            flash('Član uspješno arhiviran', 'success')
+
+    return "1"
+
+
+@members_bp.route('/erase/<member_id>', methods=['POST'])
+@admin_required
+@login_required
+def erase_member(member_id):
+
+    if request.method == 'POST':
+
+        db = DatabaseController()
+        if not db.entry_exists(DatabaseTables.CLAN, member_id):
+            error = 'Neuspješno brisanje. Zapis ne postoji u bazi.'
+            flash(error, 'danger')
+        else:
             db.remove_entry(DatabaseTables.CLAN, member_id)
-            flash('Član uspješno izbrisan', 'success')
+            flash('Član uspješno arhiviran', 'success')
+
+    return "1"
+
+
+@members_bp.route('/archive', methods=['GET'])
+@login_required
+def archive():
+    db = DatabaseController()
+    all_archived_members = db.get_row(DatabaseTables.CLAN, 'aktivan', 0, return_all=True)
+    members_list = {}
+    for member in sorted(all_archived_members, key=lambda x: x[10]):
+        members_list[member[0]] = (member[1], member[2], member[3], member[7],
+                                   member[8], member[9], member[10], member[-1])
+
+    return render_template('/members/archive.html', members_list=members_list)
+
+
+@members_bp.route('/<member_id>/activate', methods=['POST'])
+@login_required
+@savjetnik_required
+def activate(member_id=None):
+    if request.method == 'POST':
+        db = DatabaseController()
+        db.activate_member(member_id)
+        flash('Član je uspješno aktiviran', 'success')
+
+        all_archived_members = db.get_row(DatabaseTables.CLAN, 'aktivan', 0, return_all=True)
+        members_list = {}
+        for member in all_archived_members:
+            members_list[member[0]] = (member[1], member[2], member[3], member[7],
+                                       member[8], member[9], member[10], member[-1])
 
     return "1"
